refactor: 기술 면접 답변 패키지 구조 변경
