+++ backend/src/main/java/com/java/NBE4_5_1_7/domain/interview/controller/InterviewController.java
@@ -1,16 +1,28 @@
 package com.java.NBE4_5_1_7.domain.interview.controller;
 
+import java.util.List;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 import com.java.NBE4_5_1_7.domain.interview.entity.InterviewCategory;
+import com.java.NBE4_5_1_7.domain.interview.entity.dto.request.InterviewCommentDetailDto;
 import com.java.NBE4_5_1_7.domain.interview.entity.dto.request.KeywordContentRequestDto;
 import com.java.NBE4_5_1_7.domain.interview.entity.dto.request.RandomRequestDto;
 import com.java.NBE4_5_1_7.domain.interview.entity.dto.response.InterviewResponseDto;
 import com.java.NBE4_5_1_7.domain.interview.entity.dto.response.RandomResponseDto;
+import com.java.NBE4_5_1_7.domain.interview.service.InterviewCommentService;
 import com.java.NBE4_5_1_7.domain.interview.service.InterviewService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
+import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequiredArgsConstructor
@@ -52,5 +64,46 @@ public class InterviewController {
     @PostMapping("/keyword/content")
     public ResponseEntity<List<Long>> keywordContentId(@RequestBody KeywordContentRequestDto keywordContentRequestDto) {
         return ResponseEntity.ok(service.keywordHeadQuestion(keywordContentRequestDto));
+    }
+
+    @RestController
+    @RequiredArgsConstructor
+    @RequestMapping("/api/v1/interview-comments")
+    public static class InterviewCommentController {
+
+        private final InterviewCommentService interviewCommentService;
+
+        @PostMapping
+        public ResponseEntity<InterviewCommentDetailDto> createComment(
+            @RequestBody InterviewCommentDetailDto newDto) {
+            InterviewCommentDetailDto createdComment = interviewCommentService.createComment(newDto);
+            return ResponseEntity.status(HttpStatus.CREATED).body(createdComment);
+        }
+
+        @GetMapping
+        public ResponseEntity<List<InterviewCommentDetailDto>> all() {
+            List<InterviewCommentDetailDto> comments = interviewCommentService.getAllComments();
+            return ResponseEntity.ok(comments);
+        }
+
+        @GetMapping("/{commentId}")
+        public ResponseEntity<InterviewCommentDetailDto> getCommentById(@PathVariable Long commentId) {
+            InterviewCommentDetailDto comment = interviewCommentService.getCommentById(commentId);
+            return ResponseEntity.ok(comment);
+        }
+
+        @PatchMapping("/{commentId}")
+        public ResponseEntity<InterviewCommentDetailDto> updateComment(
+            @PathVariable("commentId") Long commentId,
+            @RequestBody InterviewCommentDetailDto updatedDto) {
+            InterviewCommentDetailDto updatedComment = interviewCommentService.updateComment(commentId, updatedDto);
+            return ResponseEntity.ok(updatedComment);
+        }
+
+        @DeleteMapping("/{commentId}")
+        public ResponseEntity<String> deleteComment(@PathVariable Long commentId) {
+            interviewCommentService.deleteComment(commentId);
+            return ResponseEntity.ok("답변이 삭제되었습니다.");
+        }
     }
 }
